migrate TestCalculateRewardsBasic to simapp in distribution
