+++ x/distribution/keeper/allocation_test.go
@@ -132,8 +132,6 @@ func TestAllocateTokensToManyValidators(t *testing.T) {
 }
 
 func TestAllocateTokensTruncation(t *testing.T) {
-	//communityTax := sdk.NewDec(0)
-
 	app := simapp.Setup(false)
 	ctx := app.BaseApp.NewContext(false, abci.Header{})
